experiment v0.0.1 Experiment behaviour

This module injects the lab config for working with labs. This also has the methods for creating and running lab experiments.

Example

defmodule Controller do
  use Experiment

  def main do
    lab("Test my new experimental function")
    |> experiment(&experimental_function/0)
    |> control(&control_function/0)
    |> perform_experiment
  end

  def control_function do
    {:ok, :foo}
  end

  def experimental_function do
    {:ok, :bar}
  end
end

Functions

compare_tests(control, candidate)

Specs

compare_tests(any, any) :: boolean

Callbacks

control(arg0, list)

Specs

control(Experiment.Lab.t, (... -> any)) :: Experiment.Lab.t

Adds the control in which all experiments will be compared to.

The result of this function will be the result of the experiment.

experiment(arg0, list)

Specs

experiment(Experiment.Lab.t, (... -> any)) :: Experiment.Lab.t

Adds a new experiment to the lab.

experiment(arg0, arg1, list)

Specs

experiment(Experiment.Lab.t, String.t, (... -> any)) :: Experiment.Lab.t

Adds a new experiment to the lab with the given name.

lab(arg0)

Specs

lab(String.t) :: Experiment.Lab.t

Returns a new lab experiment with the given name.

perform_experiment(arg0)

Specs

perform_experiment(Experiment.Lab.t) :: any

Runs the lab experiment, returning the result of the control.