experiment v0.0.2 Experiment.Lab

This module defines the Experiment.Lab struct and the main functions for working with the lab. This struct holds all of the experimental functions, as well as the control functions.

Types

t :: %Experiment.Lab{adapter: module, compare: Experiment.Test.t, control: function | nil, experiment_count: integer, experiments: list | nil, name: String.t}