experiment v0.0.1 Experiment

This module injects the lab config for working with labs. This also has the methods for creating and running lab experiments.

Example

defmodule Controller do

  def main do
    Experiment.new("Test my new experimental function")
    |> Experiment.experiment(&experimental_function/0)
    |> Experiment.control(&control_function/0)
    |> Experiment.perform_experiment
  end

  def control_function do
    {:ok, :foo}
  end

  def experimental_function do
    {:ok, :bar}
  end
end

Functions

compare(lab, func)

Specs

compare(Experiment.Lab.t, (... -> any)) :: Experiment.Lab.t

Overrides the default compare function on the lab.

compare_tests(control, candidate)

Specs

compare_tests(any, any) :: boolean

Default comparator. Does == against the control and candidate.

control(lab, func)

Specs

control(Experiment.Lab.t, (... -> any)) :: Experiment.Lab.t

Adds the control in which all experiments will be compared to.

The result of this function will be the result of the experiment.

new(name)

Specs

new(String.t) :: Experiment.Lab.t

Returns a new lab experiment with the given name.

perform_experiment(lab)

Specs

perform_experiment(Experiment.Lab.t) :: any

Runs the lab experiment, returning the result of the control.

test(lab, func)

Specs

test(Experiment.Lab.t, (... -> any)) :: Experiment.Lab.t

Adds a new experimental test to the lab.

test(lab, func_name, func)

Specs

test(Experiment.Lab.t, String.t, (... -> any)) :: Experiment.Lab.t

Adds a new experimental test to the lab with the given name.